+++ packages/orq-rc/src/models/operations/fileupload.ts
@@ -196,7 +196,7 @@ export const FileUploadResponseBody$inboundSchema: z.ZodType<
   file_name: z.string(),
   workspace_id: z.string(),
   created: z.string().datetime({ offset: true }).default(
-    "2025-11-27T09:45:51.496Z",
+    "2025-11-27T14:02:14.828Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -229,7 +229,7 @@ export const FileUploadResponseBody$outboundSchema: z.ZodType<
   bytes: z.number(),
   fileName: z.string(),
   workspaceId: z.string(),
-  created: z.date().default(() => new Date("2025-11-27T09:45:51.496Z"))
+  created: z.date().default(() => new Date("2025-11-27T14:02:14.828Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {