+++ packages/orq-rc/src/models/operations/listdatasets.ts
@@ -211,7 +211,7 @@ export const ListDatasetsData$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2025-11-27T09:45:48.496Z",
+    "2025-11-27T14:02:11.843Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -250,7 +250,7 @@ export const ListDatasetsData$outboundSchema: z.ZodType<
   createdById: z.string().optional(),
   updatedById: z.string().optional(),
   created: z.date().transform(v => v.toISOString()).optional(),
-  updated: z.date().default(() => new Date("2025-11-27T09:45:48.496Z"))
+  updated: z.date().default(() => new Date("2025-11-27T14:02:11.843Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {