+++ packages/orq-rc/src/models/operations/listdatasources.ts
@@ -224,7 +224,7 @@ export const ListDatasourcesData$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KB2B96M8DQW2V4ARYS53KCK2"),
+  _id: z.string().default("01KB2SYKAW5BHPQKRSCNQB553F"),
   display_name: z.string(),
   description: z.string().optional(),
   status: ListDatasourcesStatus$inboundSchema,
@@ -267,7 +267,7 @@ export const ListDatasourcesData$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   ListDatasourcesData
 > = z.object({
-  id: z.string().default("01KB2B96M8DQW2V4ARYS53KCK2"),
+  id: z.string().default("01KB2SYKAW5BHPQKRSCNQB553F"),
   displayName: z.string(),
   description: z.string().optional(),
   status: ListDatasourcesStatus$outboundSchema,