+++ packages/orq-rc/src/models/operations/retrievedatasource.ts
@@ -145,7 +145,7 @@ export const RetrieveDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KB2B96MAS2C7N1720JEBCY5W"),
+  _id: z.string().default("01KB2SYKAYYZ4QX8HNGR07FPPG"),
   display_name: z.string(),
   description: z.string().optional(),
   status: RetrieveDatasourceStatus$inboundSchema,
@@ -188,7 +188,7 @@ export const RetrieveDatasourceResponseBody$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   RetrieveDatasourceResponseBody
 > = z.object({
-  id: z.string().default("01KB2B96MAS2C7N1720JEBCY5W"),
+  id: z.string().default("01KB2SYKAYYZ4QX8HNGR07FPPG"),
   displayName: z.string(),
   description: z.string().optional(),
   status: RetrieveDatasourceStatus$outboundSchema,