+++ packages/orq-rc/RELEASES.md
@@ -5439,3 +5439,13 @@ Based on:
 - [typescript v4.1.0-rc.13] packages/orq-rc
 ### Releases
 - [NPM v4.1.0-rc.13] https://www.npmjs.com/package/@orq-ai/node/v/4.1.0-rc.13 - packages/orq-rc
+
+## 2025-11-27 14:04:00
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.661.0 (2.762.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.1.0-rc.15] packages/orq-rc
+### Releases
+- [NPM v4.1.0-rc.15] https://www.npmjs.com/package/@orq-ai/node/v/4.1.0-rc.15 - packages/orq-rc